circular dependency problems on space_delete resolved
