+++ app/server/src/models/topic/topic.model.js
@@ -1,7 +1,4 @@
 const mongoose = require("mongoose");
-const SpaceModel = require("../space/space.model");
-const ResourceModel = require("../resource/resource.model");
-
 
 const topicSchema = new mongoose.Schema(
   {
@@ -69,15 +66,6 @@ const getPopulatedTopic = async (id) => {
 
 const deleteTopic = async (id) => {
   var topic = await Topic.findById(id);
-  for (var resource_temp of topic.resources) {
-    await ResourceModel.deleteResource(resource_temp);
-  }
-  var space = await SpaceModel.Space.findById(topic.space);
-  const index = space.topics.indexOf(topic_id);
-  if (index > -1) { // only splice array when item is found
-    space.topics.splice(index, 1); // 2nd parameter means remove one item only
-  }
-  await space.save();
   topic.remove();
 };
